Move TransferTask to the architecture module
