+++ raiden/transfer/architecture.py
@@ -106,6 +106,13 @@ class Event:
 
 
 @dataclass
+class TransferTask(State):
+    # TODO: When we turn these into dataclasses it would be a good time to move common attributes
+    # of all transfer tasks like the `token_network_identifier` into the common subclass
+    pass
+
+
+@dataclass
 class SendMessageEvent(Event):
     """ Marker used for events which represent off-chain protocol messages tied
     to a channel.
